remove go mod tidy
In my case with Go 1.14 this isn't working properly, because go mod tidy
is removing all dependencies which are not found in source files.
But the source files are copied only after go mod verify. So there are
no dependencies downloaded with go mod download and needs to be downloaded
during go build every time.